UI/requirements.txt
Flask
python-dotenv
streamlit
requests